Delete old UserForm.tsx
