+++ frontend/pages/admin/UserManagementPage/components/UserForm/UserForm.tsx
@@ -1,7 +1,8 @@
-import React, { Component, FormEvent } from "react";
+import React, { Component, FormEvent, useState } from "react";
+import ReactTooltip from "react-tooltip";
 import { Link } from "react-router";
 import PATHS from "router/paths";
-import { Dispatch } from "redux";
+import { useDispatch } from "react-redux";
 
 import { ITeam } from "interfaces/team";
 import { IUserFormErrors } from "interfaces/user"; // @ts-ignore
@@ -12,9 +13,17 @@ import Button from "components/buttons/Button";
 import validatePresence from "components/forms/validators/validate_presence";
 import validEmail from "components/forms/validators/valid_email"; // @ts-ignore
 import validPassword from "components/forms/validators/valid_password"; // @ts-ignore
-import InputField from "components/forms/fields/InputField"; // @ts-ignore
-import InputFieldWithIcon from "components/forms/fields/InputFieldWithIcon"; // @ts-ignore
-import Checkbox from "components/forms/fields/Checkbox"; // @ts-ignore
+// @ts-ignore
+import { renderFlash } from "redux/nodes/notifications/actions";
+// @ts-ignore
+import validPassword from "components/forms/validators/valid_password";
+import IconToolTip from "components/IconToolTip";
+// @ts-ignore
+import InputField from "components/forms/fields/InputField";
+// @ts-ignore
+import InputFieldWithIcon from "components/forms/fields/InputFieldWithIcon";
+import Checkbox from "components/forms/fields/Checkbox";
+// @ts-ignore
 import Dropdown from "components/forms/fields/Dropdown";
 import Radio from "components/forms/fields/Radio";
 import InfoBanner from "components/InfoBanner/InfoBanner";
@@ -85,229 +94,166 @@ interface ICreateUserFormProps {
   serverErrors?: { base: string; email: string }; // "server" because this form does its own client validation
   createUserErrors?: IUserFormErrors;
   editUserErrors?: IUserFormErrors;
-  dispatch: Dispatch;
 }
 
-interface ICreateUserFormState {
-  errors: IUserFormErrors;
-  formData: IFormData;
-  isGlobalUser: boolean;
-}
+const UserForm = ({
+  availableTeams,
+  onCancel,
+  onSubmit,
+  submitText,
+  defaultName,
+  defaultEmail,
+  currentUserId,
+  currentTeam,
+  isModifiedByGlobalAdmin,
+  defaultGlobalRole,
+  defaultTeamRole,
+  defaultTeams,
+  isPremiumTier,
+  smtpConfigured,
+  canUseSso,
+  isSsoEnabled,
+  isNewUser,
+  serverErrors,
+  createUserErrors,
+  editUserErrors,
+}: ICreateUserFormProps): JSX.Element => {
+  const dispatch = useDispatch();
 
-class UserForm extends Component<ICreateUserFormProps, ICreateUserFormState> {
-  constructor(props: ICreateUserFormProps) {
-    super(props);
+  const [errors, setErrors] = useState<any>(createUserErrors);
+  const [formData, setFormData] = useState<any>({
+    email: defaultEmail || "",
+    name: defaultName || "",
+    newUserType: isNewUser ? NewUserType.AdminCreated : null,
+    password: null,
+    sso_enabled: isSsoEnabled,
+    global_role: defaultGlobalRole || null,
+    teams: defaultTeams || [],
+    currentUserId: currentUserId,
+  });
 
-    console.log("UserForm.tsx props.createUserErrors", props.createUserErrors);
+  const [isGlobalUser, setIsGlobalUser] = useState<boolean>(
+    defaultGlobalRole !== null
+  );
 
-    this.state = {
-      errors: {
-        email: null,
-        name: null,
-        password: null,
-        sso_enabled: null,
-      },
-      formData: {
-        email: props.defaultEmail || "",
-        name: props.defaultName || "",
-        newUserType: props.isNewUser ? NewUserType.AdminCreated : null,
-        password: null,
-        sso_enabled: props.isSsoEnabled,
-        global_role: props.defaultGlobalRole || null,
-        teams: props.defaultTeams || [],
-        currentUserId: props.currentUserId,
-      },
-      isGlobalUser: props.defaultGlobalRole !== null,
-    };
-  }
-
-  componentWillReceiveProps(createUserErrors: any) {
-    this.setState({
-      errors: createUserErrors,
-    });
-
-    return false;
-  }
-
-  onInputChange = (formField: string): ((value: string) => void) => {
+  const onInputChange = (formField: string): ((value: string) => void) => {
     return (value: string) => {
-      const { errors, formData } = this.state;
-
-      this.setState({
-        errors: {
-          ...errors,
-          [formField]: null,
-        },
-        formData: {
-          ...formData,
-          [formField]: value,
-        },
+      setErrors({
+        ...errors,
+        [formField]: null,
+      });
+      setFormData({
+        ...formData,
+        [formField]: value,
       });
     };
   };
 
-  onCheckboxChange = (formField: string): ((evt: string) => void) => {
+  const onCheckboxChange = (formField: string): ((evt: string) => void) => {
     return (evt: string) => {
-      return this.onInputChange(formField)(evt);
+      return onInputChange(formField)(evt);
     };
   };
 
-  onRadioChange = (formField: string): ((evt: string) => void) => {
+  const onRadioChange = (formField: string): ((evt: string) => void) => {
     return (evt: string) => {
-      return this.onInputChange(formField)(evt);
+      return onInputChange(formField)(evt);
     };
   };
 
-  onIsGlobalUserChange = (value: string): void => {
-    const { formData } = this.state;
+  const onIsGlobalUserChange = (value: string): void => {
     const isGlobalUser = value === UserTeamType.GlobalUser;
-    this.setState({
-      isGlobalUser,
-      formData: {
-        ...formData,
-        global_role: isGlobalUser ? "observer" : null,
-      },
+    setIsGlobalUser(isGlobalUser);
+    setFormData({
+      ...formData,
+      global_role: isGlobalUser ? "observer" : null,
     });
   };
 
-  onGlobalUserRoleChange = (value: string): void => {
-    const { formData } = this.state;
-    this.setState({
-      formData: {
-        ...formData,
-        global_role: value,
-      },
+  const onGlobalUserRoleChange = (value: string): void => {
+    setFormData({
+      ...formData,
+      global_role: value,
     });
   };
 
-  onSelectedTeamChange = (teams: ITeam[]): void => {
-    const { formData } = this.state;
-    this.setState({
-      formData: {
-        ...formData,
-        teams,
-      },
+  const onSelectedTeamChange = (teams: ITeam[]): void => {
+    setFormData({
+      ...formData,
+      teams,
     });
   };
 
-  onTeamRoleChange = (teams: ITeam[]): void => {
-    const { formData } = this.state;
-    this.setState({
-      formData: {
-        ...formData,
-        teams,
-      },
+  const onTeamRoleChange = (teams: ITeam[]): void => {
+    setFormData({
+      ...formData,
+      teams,
     });
-  };
-
-  onFormSubmit = (evt: FormEvent): void => {
-    const { createSubmitData, validate } = this;
-    evt.preventDefault();
-    const valid = validate();
-    if (valid) {
-      const { onSubmit } = this.props;
-      return onSubmit(createSubmitData());
-    }
   };
 
   // UserForm component can be used to create a new user or edit an existing user so submitData will be assembled accordingly
-  createSubmitData = (): IFormData => {
-    const { currentUserId, isNewUser } = this.props;
-    const {
-      isGlobalUser,
-      formData: {
-        email,
-        name,
-        newUserType,
-        password,
-        sso_enabled,
-        global_role,
-        teams,
-      },
-    } = this.state;
-
-    const submitData = {
-      email,
-      name,
-      newUserType,
-      password,
-      sso_enabled,
-      currentUserId,
-    };
+  const createSubmitData = (): IFormData => {
+    const submitData = formData;
 
     if (!isNewUser) {
       delete submitData.newUserType; // this field will not be submitted when form is used to edit an existing user
     }
 
-    if (submitData.sso_enabled || newUserType === NewUserType.AdminInvited) {
+    if (
+      submitData.sso_enabled ||
+      formData.newUserType === NewUserType.AdminInvited
+    ) {
       delete submitData.password; // this field will not be submitted with the form
     }
 
     return isGlobalUser
-      ? { ...submitData, global_role, teams: [] }
-      : { ...submitData, global_role: null, teams };
+      ? { ...submitData, global_role: formData.global_role, teams: [] }
+      : { ...submitData, global_role: null, teams: formData.teams };
   };
 
-  validate = (): boolean => {
-    const {
-      errors,
-      formData: {
-        email,
-        password,
-        newUserType,
-        sso_enabled,
-        global_role,
-        teams,
-      },
-    } = this.state;
-    const { dispatch, isNewUser } = this.props;
-
-    if (!validatePresence(email)) {
-      this.setState({
-        errors: {
-          ...errors,
-          email: "Email field must be completed",
-        },
+  const validate = (): boolean => {
+    if (!validatePresence(formData.email)) {
+      setErrors({
+        ...errors,
+        email: "Email field must be completed",
       });
 
       return false;
     }
 
-    if (!validEmail(email)) {
-      this.setState({
-        errors: {
-          ...errors,
-          email: `${email} is not a valid email`,
-        },
+    if (!validEmail(formData.email)) {
+      setErrors({
+        ...errors,
+        email: `${formData.email} is not a valid email`,
       });
 
       return false;
     }
 
-    if (isNewUser && newUserType === NewUserType.AdminCreated && !sso_enabled) {
-      if (!validatePresence(password)) {
-        this.setState({
-          errors: {
-            ...errors,
-            password: "Password field must be completed",
-          },
+    if (
+      isNewUser &&
+      formData.newUserType === NewUserType.AdminCreated &&
+      !formData.sso_enabled
+    ) {
+      if (!validatePresence(formData.password)) {
+        setErrors({
+          ...errors,
+          password: "Password field must be completed",
         });
 
         return false;
       }
-      if (!validPassword(password)) {
-        this.setState({
-          errors: {
-            ...errors,
-            password: "Password must meet the criteria below",
-          },
+      if (!validPassword(formData.password)) {
+        setErrors({
+          ...errors,
+          password: "Password must meet the criteria below",
         });
 
         return false;
       }
     }
 
-    if (!global_role && !teams.length) {
+    if (!formData.global_role && !formData.teams.length) {
       dispatch(
         renderFlash("error", `Please select at least one team for this user.`)
       );
@@ -317,12 +263,15 @@ class UserForm extends Component<ICreateUserFormProps, ICreateUserFormState> {
     return true;
   };
 
-  renderGlobalRoleForm = (): JSX.Element => {
-    const { onGlobalUserRoleChange } = this;
-    const {
-      formData: { global_role },
-    } = this.state;
-    const { isPremiumTier } = this.props;
+  const onFormSubmit = (evt: FormEvent): void => {
+    evt.preventDefault();
+    const valid = validate();
+    if (valid) {
+      return onSubmit(createSubmitData());
+    }
+  };
+
+  const renderGlobalRoleForm = (): JSX.Element => {
     return (
       <>
         {isPremiumTier && (
@@ -343,7 +292,7 @@ class UserForm extends Component<ICreateUserFormProps, ICreateUserFormState> {
         )}
         <Dropdown
           label="Role"
-          value={global_role || "Observer"}
+          value={formData.global_role || "Observer"}
           className={`${baseClass}__global-role-dropdown`}
           options={globalUserRoles}
           searchable={false}
@@ -354,7 +303,7 @@ class UserForm extends Component<ICreateUserFormProps, ICreateUserFormState> {
     );
   };
 
-  renderNoTeamsMessage = (): JSX.Element => {
+  const renderNoTeamsMessage = (): JSX.Element => {
     return (
       <div>
         <p>
@@ -375,22 +324,7 @@ class UserForm extends Component<ICreateUserFormProps, ICreateUserFormState> {
     );
   };
 
-  renderTeamsForm = (): JSX.Element => {
-    const {
-      onSelectedTeamChange,
-      renderNoTeamsMessage,
-      onTeamRoleChange,
-    } = this;
-    const {
-      availableTeams,
-      isModifiedByGlobalAdmin,
-      defaultTeamRole,
-      currentTeam,
-    } = this.props;
-    const {
-      formData: { teams },
-    } = this.state;
-
+  const renderTeamsForm = (): JSX.Element => {
     return (
       <>
         {!!availableTeams.length &&
@@ -412,15 +346,15 @@ class UserForm extends Component<ICreateUserFormProps, ICreateUserFormState> {
               </InfoBanner>
               <SelectedTeamsForm
                 availableTeams={availableTeams}
-                usersCurrentTeams={teams}
+                usersCurrentTeams={formData.teams}
                 onFormChange={onSelectedTeamChange}
               />
             </>
           ) : (
             <SelectRoleForm
               label="Role"
-              currentTeam={currentTeam || teams[0]}
-              teams={teams}
+              currentTeam={currentTeam || formData.teams[0]}
+              teams={formData.teams}
               defaultTeamRole={defaultTeamRole || "observer"}
               onFormChange={onTeamRoleChange}
             />
@@ -430,60 +364,66 @@ class UserForm extends Component<ICreateUserFormProps, ICreateUserFormState> {
     );
   };
 
-  render(): JSX.Element {
-    const {
-      errors,
-      formData: { email, name, newUserType, password, sso_enabled },
-      isGlobalUser,
-    } = this.state;
-    const {
-      onCancel,
-      submitText,
-      isPremiumTier,
-      smtpConfigured,
-      canUseSso,
-      isNewUser,
-      currentTeam,
-      isModifiedByGlobalAdmin,
-      serverErrors,
-      availableTeams,
-    } = this.props;
-    const {
-      onFormSubmit,
-      onInputChange,
-      onCheckboxChange,
-      onRadioChange,
-      onIsGlobalUserChange,
-      renderGlobalRoleForm,
-      renderTeamsForm,
-    } = this;
+  if (!isPremiumTier && !isGlobalUser) {
+    console.log(
+      `Note: Fleet Free UI does not have teams options.\n
+        User ${formData.name} is already assigned to a team and cannot be reassigned without access to Fleet Premium UI.`
+    );
+  }
 
-    if (!isPremiumTier && !isGlobalUser) {
-      console.log(
-        `Note: Fleet Free UI does not have teams options.\n
-        User ${name} is already assigned to a team and cannot be reassigned without access to Fleet Premium UI.`
-      );
-    }
-
-    return (
-      <form className={baseClass} autoComplete="off">
-        {/* {baseError && <div className="form__base-error">{baseError}</div>} */}
+  return (
+    <form className={baseClass} autoComplete="off">
+      {/* {baseError && <div className="form__base-error">{baseError}</div>} */}
+      <InputFieldWithIcon
+        autofocus
+        error={errors.name}
+        name="name"
+        onChange={onInputChange("name")}
+        placeholder="Full name"
+        value={formData.name || ""}
+        inputOptions={{
+          maxLength: "80",
+        }}
+      />
+      <div
+        className="email-disabled"
+        data-tip
+        data-for="email-disabled-tooltip"
+        data-tip-disable={isNewUser || smtpConfigured}
+      >
         <InputFieldWithIcon
-          autofocus
-          error={errors.name}
-          name="name"
-          onChange={onInputChange("name")}
-          placeholder="Full name"
-          value={name || ""}
-          inputOptions={{
-            maxLength: "80",
-          }}
+          error={errors.email || serverErrors?.email}
+          name="email"
+          onChange={onInputChange("email")}
+          placeholder="Email"
+          value={formData.email || ""}
+          disabled={!isNewUser && !smtpConfigured}
         />
+      </div>
+      <ReactTooltip
+        place="bottom"
+        type="dark"
+        effect="solid"
+        id="email-disabled-tooltip"
+        backgroundColor="#3e4771"
+        data-html
+      >
+        <span className={`${baseClass}__tooltip-text`}>
+          Editing an email address requires that SMTP is <br />
+          configured in order to send a validation email. <br />
+          <br />
+          Users with Admin role can configure SMTP in
+          <br />
+          <strong>Settings &gt; Organization settings</strong>.
+        </span>
+      </ReactTooltip>
+      <div className={`${baseClass}__sso-input`}>
         <div
-          className="email-disabled"
+          className="sso-disabled"
           data-tip
-          data-for="email-disabled-tooltip"
-          data-tip-disable={isNewUser || smtpConfigured}
+          data-for="sso-disabled-tooltip"
+          data-tip-disable={canUseSso}
+          data-offset="{'top': 25, 'left': 100}"
         >
           <InputFieldWithIcon
             label="Email"
@@ -491,7 +431,7 @@ class UserForm extends Component<ICreateUserFormProps, ICreateUserFormState> {
             name="email"
             onChange={onInputChange("email")}
             placeholder="Email"
-            value={email || ""}
+            value={formData.email || ""}
             disabled={!isNewUser && !smtpConfigured}
             tooltip={
               "\
@@ -506,7 +446,7 @@ class UserForm extends Component<ICreateUserFormProps, ICreateUserFormState> {
           <Checkbox
             name="sso_enabled"
             onChange={onCheckboxChange("sso_enabled")}
-            value={canUseSso && sso_enabled}
+            value={canUseSso && formData.sso_enabled}
             disabled={!canUseSso}
             wrapperClassName={`${baseClass}__invite-admin`}
             tooltip={`
@@ -521,51 +461,54 @@ class UserForm extends Component<ICreateUserFormProps, ICreateUserFormState> {
             Password authentication will be disabled for this user.
           </p>
         </div>
-        {isNewUser && (
-          <div className={`${baseClass}__new-user-container`}>
-            <div className={`${baseClass}__new-user-radios`}>
-              {isModifiedByGlobalAdmin ? (
-                <>
-                  <Radio
-                    className={`${baseClass}__radio-input`}
-                    label={"Create user"}
-                    id={"create-user"}
-                    checked={newUserType !== NewUserType.AdminInvited}
-                    value={NewUserType.AdminCreated}
-                    name={"newUserType"}
-                    onChange={onRadioChange("newUserType")}
-                  />
-                  <Radio
-                    className={`${baseClass}__radio-input`}
-                    label={"Invite user"}
-                    id={"invite-user"}
-                    disabled={!smtpConfigured}
-                    checked={newUserType === NewUserType.AdminInvited}
-                    value={NewUserType.AdminInvited}
-                    name={"newUserType"}
-                    onChange={onRadioChange("newUserType")}
-                    tooltip={
-                      smtpConfigured
-                        ? ""
-                        : `
+      </div>
+      {isNewUser && (
+        <div className={`${baseClass}__new-user-container`}>
+          <div className={`${baseClass}__new-user-radios`}>
+            {isModifiedByGlobalAdmin ? (
+              <>
+                <Radio
+                  className={`${baseClass}__radio-input`}
+                  label={"Invite user"}
+                  id={"invite-user"}
+                  disabled={!smtpConfigured}
+                  checked={formData.newUserType === NewUserType.AdminInvited}
+                  value={NewUserType.AdminInvited}
+                  name={"newUserType"}
+                  onChange={onRadioChange("newUserType")}
+                />
+                <Radio
+                  className={`${baseClass}__radio-input`}
+                  label={"Invite user"}
+                  id={"invite-user"}
+                  disabled={!smtpConfigured}
+                  checked={formData.newUserType === NewUserType.AdminInvited}
+                  value={NewUserType.AdminInvited}
+                  name={"newUserType"}
+                  onChange={onRadioChange("newUserType")}
+                  tooltip={
+                    smtpConfigured
+                      ? ""
+                      : `
                       The &quot;Invite user&quot; feature requires that SMTP
                       is configured in order to send invitation emails.
                       <br /><br />
                       SMTP can be configured in Settings &gt; Organization settings.
                     `
-                    }
-                  />
-                </>
-              ) : (
-                <input
-                  type="hidden"
-                  id={"create-user"}
-                  value={NewUserType.AdminCreated}
-                  name={"newUserType"}
+                  }
                 />
-              )}
-            </div>
-            {newUserType !== NewUserType.AdminInvited && !sso_enabled && (
+              </>
+            ) : (
+              <input
+                type="hidden"
+                id={"create-user"}
+                value={NewUserType.AdminCreated}
+                name={"newUserType"}
+              />
+            )}
+          </div>
+          {formData.newUserType !== NewUserType.AdminInvited &&
+            !formData.sso_enabled && (
               <>
                 <div className={`${baseClass}__password`}>
                   <InputField
@@ -574,7 +517,7 @@ class UserForm extends Component<ICreateUserFormProps, ICreateUserFormState> {
                     name="password"
                     onChange={onInputChange("password")}
                     placeholder="Password"
-                    value={password || ""}
+                    value={formData.password || ""}
                     type="password"
                     hint={[
                       "Must include 7 characters, at least 1 number (e.g. 0 - 9), and at least 1 symbol (e.g. &*#)",
@@ -588,67 +531,66 @@ class UserForm extends Component<ICreateUserFormProps, ICreateUserFormState> {
                 </div>
               </>
             )}
-          </div>
-        )}
-        {isPremiumTier && (
-          <div className={`${baseClass}__selected-teams-container`}>
-            <div className={`${baseClass}__team-radios`}>
-              <p className={`${baseClass}__label`}>Team</p>
-              {isModifiedByGlobalAdmin ? (
-                <>
-                  <Radio
-                    className={`${baseClass}__radio-input`}
-                    label={"Global user"}
-                    id={"global-user"}
-                    checked={isGlobalUser}
-                    value={UserTeamType.GlobalUser}
-                    name={"userTeamType"}
-                    onChange={onIsGlobalUserChange}
-                  />
-                  <Radio
-                    className={`${baseClass}__radio-input`}
-                    label={"Assign teams"}
-                    id={"assign-teams"}
-                    checked={!isGlobalUser}
-                    value={UserTeamType.AssignTeams}
-                    name={"userTeamType"}
-                    onChange={onIsGlobalUserChange}
-                    disabled={!availableTeams.length}
-                  />
-                </>
-              ) : (
-                <p className="current-team">
-                  {currentTeam ? currentTeam.name : ""}
-                </p>
-              )}
-            </div>
-            <div className={`${baseClass}__teams-form-container`}>
-              {isGlobalUser ? renderGlobalRoleForm() : renderTeamsForm()}
-            </div>
-          </div>
-        )}
-        {!isPremiumTier && renderGlobalRoleForm()}
-
-        <div className={`${baseClass}__btn-wrap`}>
-          <Button
-            className={`${baseClass}__btn`}
-            type="button"
-            variant="brand"
-            onClick={onFormSubmit}
-          >
-            {submitText}
-          </Button>
-          <Button
-            className={`${baseClass}__btn`}
-            onClick={onCancel}
-            variant="inverse"
-          >
-            Cancel
-          </Button>
         </div>
-      </form>
-    );
-  }
-}
+      )}
+      {isPremiumTier && (
+        <div className={`${baseClass}__selected-teams-container`}>
+          <div className={`${baseClass}__team-radios`}>
+            <p className={`${baseClass}__label`}>Team</p>
+            {isModifiedByGlobalAdmin ? (
+              <>
+                <Radio
+                  className={`${baseClass}__radio-input`}
+                  label={"Global user"}
+                  id={"global-user"}
+                  checked={isGlobalUser}
+                  value={UserTeamType.GlobalUser}
+                  name={"userTeamType"}
+                  onChange={onIsGlobalUserChange}
+                />
+                <Radio
+                  className={`${baseClass}__radio-input`}
+                  label={"Assign teams"}
+                  id={"assign-teams"}
+                  checked={!isGlobalUser}
+                  value={UserTeamType.AssignTeams}
+                  name={"userTeamType"}
+                  onChange={onIsGlobalUserChange}
+                  disabled={!availableTeams.length}
+                />
+              </>
+            ) : (
+              <p className="current-team">
+                {currentTeam ? currentTeam.name : ""}
+              </p>
+            )}
+          </div>
+          <div className={`${baseClass}__teams-form-container`}>
+            {isGlobalUser ? renderGlobalRoleForm() : renderTeamsForm()}
+          </div>
+        </div>
+      )}
+      {!isPremiumTier && renderGlobalRoleForm()}
+
+      <div className={`${baseClass}__btn-wrap`}>
+        <Button
+          className={`${baseClass}__btn`}
+          type="button"
+          variant="brand"
+          onClick={onFormSubmit}
+        >
+          {submitText}
+        </Button>
+        <Button
+          className={`${baseClass}__btn`}
+          onClick={onCancel}
+          variant="inverse"
+        >
+          Cancel
+        </Button>
+      </div>
+    </form>
+  );
+};
 
 export default UserForm;
