+++ frontend/pages/admin/UserManagementPage/components/UserForm/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./UserFormFunctional";
+export { default } from "./UserForm";
